reorganization - some cleaning
