+++ MessageService/src/controllers/sendMessage.js
@@ -15,7 +15,6 @@ module.exports = function(req, res) {
 
     if (current_credit > 0) {
       const postOptions = {
-        // host: "exercise4_messageapp_1",
         //host: "messageapp",
         host: "localhost",
         port: 3000,
